refactor(apps): migrate all apps to v2.3.0 effects-first API
## Apps Migrated
- todo-app: createApp({ schema, effects })
- logistics-app: createApp({ schema, effects })
- taskflow: createApp({ schema, effects })

## Framework Changes
- system-get.ts: Support both read mode (path) and generate mode (key)
  - Read mode: { path, target } - read from snapshot
  - Generate mode: { key, into } - generate uuid/timestamp for $system.*
- Remove unnecessary RESERVED_EFFECT_PREFIX validation
- Keep RESERVED_EFFECT_TYPE validation (users can't override system.get)

## Removed
- @manifesto-ai/host dependency from all apps
- Legacy handlers.ts files (replaced with effects.ts)
- Legacy host-adapter.ts, array.ts, system.ts from taskflow

## Renamed (cleanup from previous session)
- v2-executor.ts → executor.ts
- v2-initializer.ts → initializer.ts
- _v2* prefixed fields → cleaner names
