--- apps/taskflow/src/manifesto/effects/system.ts
@@ -1,81 +1,0 @@
-/**
- * System Effect Handlers
- *
- * Implements Host effect handlers for system value acquisition.
- * These are generated by the compiler's lowering pass for $system.* references.
- *
- * Effect Pattern:
- *   effect system.get({ key: "uuid", into: "__sys__action_uuid_value" })
- */
-
-import type { EffectHandler } from "@manifesto-ai/host";
-import type { Patch } from "@manifesto-ai/core";
-
-/**
- * UUID generator using crypto
- */
-function generateUUID(): string {
-  if (typeof crypto !== "undefined" && crypto.randomUUID) {
-    return crypto.randomUUID();
-  }
-  // Fallback for environments without crypto.randomUUID
-  return "xxxxxxxx-xxxx-4xxx-yxxx-xxxxxxxxxxxx".replace(/[xy]/g, (c) => {
-    const r = (Math.random() * 16) | 0;
-    const v = c === "x" ? r : (r & 0x3) | 0x8;
-    return v.toString(16);
-  });
-}
-
-/**
- * system.get effect handler
- *
- * Retrieves system values like uuid, timestamp, etc.
- */
-export const systemGetHandler: EffectHandler = async (
-  _type: string,
-  params: Record<string, unknown>
-): Promise<Patch[]> => {
-  const { key, into } = params;
-
-  if (typeof into !== "string") {
-    console.error("system.get: 'into' must be a string path");
-    return [];
-  }
-
-  if (typeof key !== "string") {
-    console.error("system.get: 'key' must be a string");
-    return [];
-  }
-
-  let value: unknown;
-
-  switch (key) {
-    case "uuid":
-      value = generateUUID();
-      break;
-
-    case "timestamp":
-    case "time.now":
-      value = Date.now();
-      break;
-
-    case "isoTimestamp":
-      value = new Date().toISOString();
-      break;
-
-    default:
-      console.warn(`system.get: unknown key "${key}"`);
-      value = null;
-  }
-
-  return [{ op: "set", path: into, value }];
-};
-
-/**
- * Register all system effect handlers
- */
-export function registerSystemEffects(
-  register: (type: string, handler: EffectHandler) => void
-): void {
-  register("system.get", systemGetHandler);
-}
